update readme to not use the git url handler
the git url (git://) is not secure and the port is often blocked in large enterprises (I'm not even sure github supports it anymore).  The https and ssh urls are safer and not usually blocked (ssh is sometimes).  The https url has the advantage/disadvantage of honoring any https proxies.